Use default values for config_accessor
Hi,
I came across this file and stumbled upon your comment. A look in the AS::Configurable module revealed a way to set default values for config_accessor. I thought i just share this with you. This is not tested (I did not even clone the repo). This PR is to be considered as a hint, nothing more. Since  I don't understand the code that well, I did not know how to make `param_name` have a default value.
Hope this helps :)